move dataclass_repr in torchgen
