+++ torchgen/utils.py
@@ -4,7 +4,12 @@ import hashlib
 import os
 import re
 import textwrap
+import sys
 from argparse import Namespace
+from dataclasses import (
+    fields,
+    is_dataclass,
+)
 from typing import (
     Tuple,
     List,
@@ -287,3 +292,69 @@ def make_file_manager(
     return FileManager(
         install_dir=install_dir, template_dir=template_dir, dry_run=options.dry_run
     )
+
+
+# Helper function to create a pretty representation for dataclasses
+def dataclass_repr(obj, indent=0, width=80) -> str:
+    # built-in pprint module support dataclasses from python 3.10
+    if sys.version_info >= (3, 10):
+        from pprint import pformat
+        return pformat(obj, indent, width)
+
+    return _pformat(obj, indent=indent, width=width)
+
+def _pformat(obj, indent, width, curr_indent=0) -> str:
+    assert is_dataclass(obj), f"obj should be a dataclass, received: {type(obj)}"
+
+    class_name = obj.__class__.__name__
+    # update current indentation level with class name
+    curr_indent += len(class_name) + 1
+
+    fields_list = [(f.name, getattr(obj, f.name)) for f in fields(obj) if f.repr]
+
+    fields_str = []
+    for name, attr in fields_list:
+        # update the current indent level with the field name
+        # dict, list, set and tuple also add indent as done in pprint
+        _curr_indent = curr_indent + len(name) + 1
+        if is_dataclass(attr):
+            str_repr = _pformat(attr, indent, width, _curr_indent)
+        elif isinstance(attr, dict):
+            str_repr = _format_dict(attr, indent, width, _curr_indent)
+        elif isinstance(attr, (list, set, tuple)):
+            str_repr = _format_list(attr, indent, width, _curr_indent)
+        else:
+            str_repr = repr(attr)
+
+        fields_str.append(f"{name}={str_repr}")
+
+    indent_str = curr_indent * " "
+    body = f",\n{indent_str}".join(fields_str)
+    return f"{class_name}({body})"
+
+def _format_dict(attr, indent, width, curr_indent) -> str:
+    curr_indent += (indent + 3)
+    dict_repr = []
+    for k, v in attr.items():
+        k_repr = repr(k)
+        v_str = _pformat(v, indent, width, curr_indent + len(k_repr)) if is_dataclass(v) else repr(v)
+        dict_repr.append(f"{k_repr}: {v_str}")
+
+    return _format(dict_repr, indent, width, curr_indent, "{", "}")
+
+def _format_list(attr, indent, width, curr_indent) -> str:
+    curr_indent += (indent + 1)
+    list_repr = [_pformat(l, indent, width, curr_indent) if is_dataclass(l) else repr(l) for l in attr]
+    start, end = ("[", "]") if isinstance(attr, list) else ("(", ")")
+    return _format(list_repr, indent, width, curr_indent, start, end)
+
+def _format(fields_str, indent, width, curr_indent, start, end) -> str:
+    delimiter, curr_indent_str = "", ""
+    # if it exceed the max width then we place one element per line
+    if len(repr(fields_str)) >= width:
+        delimiter = "\n"
+        curr_indent_str = " " * curr_indent
+
+    indent_str = " " * indent
+    fields_str = f", {delimiter}{curr_indent_str}".join(fields_str)
+    return f"{start}{indent_str}{fields_str}{end}"
